setup gitignore
